requirements.txt
googlesearch-python 
cryptography
beautifulsoup4